pyproject.toml
[tool.poetry]
name = "network-live"
version = "0.1.0"
description = ""
authors = ["Ramil Albakov <ralbakov10@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.21.0"
cx-Oracle = "^8.3.0"
enm-client-scripting = {path = "enm_client_scripting-1.20.1-py2.py3-none-any.whl"}

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.16.1"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
ipython = "^8.4.0"

[tool.poetry.scripts]
update-enm-iot = "network_live.scripts.update_enm_iot:main"
update-enm-nr = "network_live.scripts.update_enm_nr:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
